Rename incentive service to provider service
