+++ .github/integration-test/start-services.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-JARS=(info store incentive dsprotection)
+JARS=(info store provider)
 
 # shellcheck disable=SC2164
 pushd services
