--- services/incentive/src/main/java/org/cryptimeleon/incentive/services/incentive/SpringConfig.java
@@ -1,9 +1,0 @@
-package org.cryptimeleon.incentive.services.incentive;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@Configuration
-@EnableScheduling
-public class SpringConfig {
-}
