--- services/incentive/src/main/java/org/cryptimeleon/incentive/services/incentive/error/BasketAlreadyPaidException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentive.services.incentive.error;
-
-public class BasketAlreadyPaidException extends IncentiveServiceException {
-    public BasketAlreadyPaidException() {
-        super("");
-    }
-}
