--- services/incentive/src/main/java/org/cryptimeleon/incentive/services/incentive/error/IncentiveServiceException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentive.services.incentive.error;
-
-public class IncentiveServiceException extends RuntimeException {
-    public IncentiveServiceException(String message) {
-        super(message);
-    }
-}
