--- services/incentive/src/main/java/org/cryptimeleon/incentive/services/incentive/error/OnlineDoubleSpendingException.java
@@ -1,7 +1,0 @@
-package org.cryptimeleon.incentive.services.incentive.error;
-
-public class OnlineDoubleSpendingException extends IncentiveServiceException {
-    public OnlineDoubleSpendingException() {
-        super("");
-    }
-}
