+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/ProviderApplication.java
@@ -1,0 +1,23 @@
+package org.cryptimeleon.incentive.services.provider;
+
+import org.cryptimeleon.incentive.client.InfoClient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+@SpringBootApplication
+public class ProviderApplication {
+
+    @Value("${info-service.url}")
+    private String infoServiceUrl;
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProviderApplication.class, args);
+    }
+
+    @Bean
+    InfoClient infoClient() {
+        return new InfoClient(infoServiceUrl);
+    }
+}
