+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/SpringConfig.java
@@ -1,0 +1,9 @@
+package org.cryptimeleon.incentive.services.provider;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class SpringConfig {
+}
