+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/error/BasketAlreadyPaidException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.services.provider.error;
+
+public class BasketAlreadyPaidException extends IncentiveServiceException {
+    public BasketAlreadyPaidException() {
+        super("");
+    }
+}
