+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/error/BasketNotPaidException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.services.provider.error;
+
+public class BasketNotPaidException extends IncentiveServiceException {
+    public BasketNotPaidException() {
+        super("Basket must be paid to retrieve the results of Earn/Spend!");
+    }
+}
