+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/error/IncentiveServiceException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.services.provider.error;
+
+public class IncentiveServiceException extends RuntimeException {
+    public IncentiveServiceException(String message) {
+        super(message);
+    }
+}
