+++ services/provider/src/main/java/org/cryptimeleon/incentive/services/provider/error/OnlineDoubleSpendingException.java
@@ -1,0 +1,7 @@
+package org.cryptimeleon.incentive.services.provider.error;
+
+public class OnlineDoubleSpendingException extends IncentiveServiceException {
+    public OnlineDoubleSpendingException() {
+        super("");
+    }
+}
