+++ services/settings.gradle
@@ -5,5 +5,5 @@ includeBuild '../promotion'
 include ':store'
 include ':client'
 include ':info'
-include ':incentive'
+include ':provider'
 include ':bootstrap'
